requirements.txt
asn1crypto==1.3.0
cached-property==1.5.1
pycryptodome==3.9.7
python-pkcs11==0.6.0
